app/server/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and caching
redis==5.0.1
aioredis==2.0.1

# Celery
celery==5.3.4
celery[redis]==5.3.4

# Telegram Bot
aiogram==3.2.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# AI/LLM
openai==1.3.7

# Storage
boto3==1.34.0
minio==7.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and logging
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
phonenumbers==8.13.26

